+++ src/main/java/de/uib/configed/gui/NewClientDialog.java
@@ -110,7 +110,7 @@ public final class NewClientDialog extends FGeneralDialog {
 
 	public void setDefaultValues() {
 		jCheckWan.setSelected(persistenceController.getConfigDataService()
-				.isInstallByShutdownConfigured(persistenceController.getHostInfoCollections().getConfigServer()));
+				.isWanConfigured(persistenceController.getHostInfoCollections().getConfigServer()));
 		jCheckShutdownInstall.setSelected(persistenceController.getConfigDataService()
 				.isInstallByShutdownConfigured(persistenceController.getHostInfoCollections().getConfigServer()));
 	}