+++ src/main/java/de/uib/opsicommand/ConnectionErrorReporter.java
@@ -18,7 +18,6 @@ import de.uib.Main;
 import de.uib.configed.Configed;
 import de.uib.configed.ConfigedMain;
 import de.uib.configed.gui.FTextArea;
-import de.uib.opsicommand.certificate.CertificateDownloader;
 import de.uib.opsicommand.certificate.CertificateManager;
 import de.uib.utils.logging.Logging;
 import de.uib.utils.swing.FEditRecord;
@@ -111,13 +110,13 @@ public final class ConnectionErrorReporter {
 		if (choice == 1) {
 			conStat = new ConnectionState(ConnectionState.INTERRUPTED);
 		} else if (choice == 2) {
-			CertificateDownloader.downloadCertificateFile();
-			CertificateManager.saveCertificate(CertificateDownloader.getDownloadedCertificateFile());
+			CertificateManager.downloadCertificateFile();
+			CertificateManager.saveCertificate();
 			if (conStat.getState() != ConnectionState.INTERRUPTED) {
 				conStat = new ConnectionState(ConnectionState.RETRY_CONNECTION);
 			}
 		} else if (choice == 3) {
-			CertificateDownloader.downloadCertificateFile();
+			CertificateManager.downloadCertificateFile();
 			if (conStat.getState() != ConnectionState.INTERRUPTED) {
 				conStat = new ConnectionState(ConnectionState.RETRY_CONNECTION);
 			}