+++ src/main/java/de/uib/opsidatamodel/serverdata/dataservice/UserDataService.java
@@ -14,6 +14,7 @@ import java.util.Map;
 import de.uib.configed.ConfigedMain;
 import de.uib.opsicommand.AbstractPOJOExecutioner;
 import de.uib.opsicommand.OpsiMethodCall;
+import de.uib.opsicommand.ServerFacade;
 import de.uib.opsidatamodel.serverdata.CacheIdentifier;
 import de.uib.opsidatamodel.serverdata.CacheManager;
 import de.uib.opsidatamodel.serverdata.RPCMethodName;
@@ -69,8 +70,15 @@ public class UserDataService {
 		return otpSecret;
 	}
 
-	public String getOpsiCACert() {
-		OpsiMethodCall omc = new OpsiMethodCall(RPCMethodName.GET_OPSI_CA_CERT, new Object[0]);
+	public String getCACerts() {
+		RPCMethodName method;
+		if (ServerFacade.getOpsiServerVersionRetriever().isServerVersionAtLeast("4.3.18.15")) {
+			method = RPCMethodName.GET_CA_CERTS;
+		} else {
+			method = RPCMethodName.GET_OPSI_CA_CERT;
+		}
+
+		OpsiMethodCall omc = new OpsiMethodCall(method, new Object[0]);
 		return exec.getStringResult(omc);
 	}
 }